Found some spaghetti and ate it
