+++ project/app/src/main/java/com/chalmers/tda367/localfeud/ui/MainActivity.java
@@ -11,16 +11,8 @@ import android.view.View;
 
 import com.chalmers.tda367.localfeud.R;
 import com.chalmers.tda367.localfeud.data.Post;
-import com.chalmers.tda367.localfeud.util.GsonHandler;
-import com.chalmers.tda367.localfeud.util.RestClient;
+import com.chalmers.tda367.localfeud.net.ServerComm;
 import com.chalmers.tda367.localfeud.util.TagHandler;
-import com.loopj.android.http.AsyncHttpResponseHandler;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-
-import cz.msebera.android.httpclient.Header;
-import cz.msebera.android.httpclient.client.HttpResponseException;
 
 public class MainActivity extends AppCompatActivity implements PostAdapter.AdapterCallback {
 
@@ -28,110 +20,13 @@ public class MainActivity extends AppCompatActivity implements PostAdapter.Adapt
     private PostAdapter postAdapter;
     private FloatingActionButton createNewFab;
 
-    //    TEMP STUFF
-    private RestClient restClient;
-
-    private ArrayList<Post> dummyPostList;
-    private Post dummyPost;
-    private String dummyJsonPost = "// 20160413151533\n" +
-            "// http://api-local.ericwenn.se/posts/1/\n" +
-            "\n" +
-            "{\n" +
-            "  \"id\": \"1\",\n" +
-            "  \"location\": {\n" +
-            "    \"latitude\": 32.1231,\n" +
-            "    \"longitude\": 13.123123,\n" +
-            "    \"distance\": 7\n" +
-            "  },\n" +
-            "  \"user\": {\n" +
-            "    \"id\": 2,\n" +
-            "    \"firstname\": \"Krune\",\n" +
-            "    \"lastname\": \"Nilsson\",\n" +
-            "    \"href\": \"http://localhost/local-feud_backend/src/users/2/\"\n" +
-            "  },\n" +
-            "  \"reach\": 5,\n" +
-            "  \"content\": {\n" +
-            "    \"type\": \"text\",\n" +
-            "    \"text\": \"Lorem ipsum dolorem.\"\n" +
-            "  },\n" +
-            "  \"date_posted\": \"2016-04-13T15:15:33+02:00\",\n" +
-            "  \"is_deleted\": false,\n" +
-            "  \"number_of_comments\": 5,\n" +
-            "  \"number_of_likes\": 10\n" +
-            "}";
-    private String dummyJsonPostList = "// 20160413150946\n" +
-            "// http://api-local.ericwenn.se/posts/\n" +
-            "\n" +
-            "[\n" +
-            "  {\n" +
-            "    \"id\": \"1\",\n" +
-            "    \"reach\": \"500\",\n" +
-            "    \"date_posted\": \"2016-04-12T13:37:18+02:00\",\n" +
-            "    \"location\": {\n" +
-            "      \"distance\": 275780.74151391\n" +
-            "    },\n" +
-            "    \"user\": {\n" +
-            "      \"id\": \"1\",\n" +
-            "      \"href\": \"/users/1/\"\n" +
-            "    },\n" +
-            "    \"number_of_comments\": 4,\n" +
-            "    \"number_of_likes\": 10,\n" +
-            "    \"content\": {\n" +
-            "      \"type\": \"text\",\n" +
-            "      \"text\": \"Lorem ipsum\"\n" +
-            "    },\n" +
-            "    \"href\": \"/posts/1/\"\n" +
-            "  },\n" +
-            "  {\n" +
-            "    \"id\": \"2\",\n" +
-            "    \"reach\": \"1000\",\n" +
-            "    \"date_posted\": \"2016-04-12T13:42:29+02:00\",\n" +
-            "    \"location\": {\n" +
-            "      \"distance\": 276875.61419316\n" +
-            "    },\n" +
-            "    \"user\": {\n" +
-            "      \"id\": \"1\",\n" +
-            "      \"href\": \"/users/1/\"\n" +
-            "    },\n" +
-            "    \"number_of_comments\": 4,\n" +
-            "    \"number_of_likes\": 10,\n" +
-            "    \"content\": {\n" +
-            "      \"type\": \"text\",\n" +
-            "      \"text\": \"Lorem ipsum\"\n" +
-            "    },\n" +
-            "    \"href\": \"/posts/2/\"\n" +
-            "  }\n" +
-            "]";
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        dummyPost = GsonHandler.getInstance().toPost(dummyJsonPost);
-        dummyPostList = GsonHandler.getInstance().toPostList(dummyJsonPostList);
         initViews();
 
-//        SPAGHETTI-CODE INCOMING
-        restClient = new RestClient();
-        HashMap<String, String> tempMap = new HashMap<>();
-        tempMap.put("test", "test");
-        try {
-            restClient.get("posts/", tempMap, new AsyncHttpResponseHandler() {
-                @Override
-                public void onSuccess(int statusCode, Header[] headers, byte[] responseBody) {
-                    ArrayList<Post> posts = GsonHandler.getInstance().toPostList(new String(responseBody));
-                    postAdapter.addPostListToAdapter(posts);
-                    Log.d(TagHandler.MAIN_TAG, "onSuccess. Posts: " + posts.size());
-                }
-
-                @Override
-                public void onFailure(int statusCode, Header[] headers, byte[] responseBody, Throwable error) {
-                    Log.e(TagHandler.MAIN_TAG, error.getMessage());
-                }
-            });
-        } catch (HttpResponseException e) {
-            Log.e(TagHandler.MAIN_TAG, e.getMessage());
-        }
+        ServerComm.updatePostFeed(postAdapter);
     }
 
     private void initViews() {
@@ -154,9 +49,6 @@ public class MainActivity extends AppCompatActivity implements PostAdapter.Adapt
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         postAdapter = new PostAdapter(this);
         recyclerView.setAdapter(postAdapter);
-//        for (Post post : dummyPostList) {
-//            postAdapter.addPostToAdapter(post);
-//        }
     }
 
     @Override
